Unify VlcConfig integer fields on NonZeroU64
`max_v8_heap_size_mb`, `max_ephemeral_workers`, and `num_workers`
were `NonZeroUsize`; the rest were `NonZeroU64`. The mixed types
matched their consumer APIs (V8's `usize`, `Vec` capacity) but the
heterogeneity has no real benefit at the public-API level. Convert
to `usize` at the consumption sites — three call sites, all infallible
in practice.

Python's `extract_positive_usize` collapses into `extract_positive_u64`.

+++ vl-convert/src/cli_types.rs
@@ -143,7 +143,7 @@ pub(crate) struct Cli {
 
     /// Maximum V8 heap size per worker in megabytes [default: 0 = no limit]
     #[arg(long, global = true)]
-    pub(crate) max_v8_heap_size_mb: Option<usize>,
+    pub(crate) max_v8_heap_size_mb: Option<u64>,
 
     /// Maximum V8 execution time in seconds [default: 0 = no limit]
     #[arg(long, global = true)]
